BaseDtos

+++ src/IczpNet.Organization.Application.Contracts/Functionals/Dtos/FunctionalDto.cs
@@ -9,7 +9,7 @@ namespace IczpNet.Organization.Functionals.Dtos
         /// <summary>
         /// 排序（越大越前面） DESC
         /// </summary>
-        public virtual long Sorting { get; set; }
+        public virtual double Sorting { get; set; }
         /// <summary>
         /// 说明
         /// </summary>
